fix(SC-922), Fix setter function for array table element in group
Fix setter function for array table element which is a
child of group to call the setter function with correct
number of parameters in SIL.

Previously for grid support tables, the *setNumActivePoints
function was called with only 1 parameter but the function
takes 4. The first parameter is the selected grid support curve
but the function was called with num active points in its
place and it would cause overindexing if the number of
active points was larger than 3.