fix(SC-2749): Sort children without copying
The old one sorted the top level correctly but the signal within mux
were lost. The new version retains children but the tree view does not
get updated. I can't seem to figure out how to get the tree updated
with small effort so this will have to do for now